first commit from terminal
