+++ src/util/TaskProgress.java
@@ -1,0 +1,5 @@
+package util;
+
+public enum TaskProgress {
+    NEW, IN_PROGRESS, DONE
+}
