+++ src/util/historyManager/HistoryManager.java
@@ -1,0 +1,4 @@
+package util.historyManager;
+
+public class HistoryManager {
+}
