+++ src/util/taskManager/Managers.java
@@ -1,0 +1,8 @@
+package util.taskManager;
+
+public class Managers <T extends TaskManager> {
+
+    public InMemoryTaskManager getDefault() {
+        return new InMemoryTaskManager();
+    }
+}
