+++ tasks.txt
@@ -1,3 +1,1 @@
-1:qwe`as`false`false`1
-2:aaaa`zzzz`true`true`2`{cccc=Done}
-3:ag`vv`false`true`3`{q=In_Progress, fdg=Done, ghj=Done, yu=Done}
+1:qwe`sd`true`false`1
